+++ scripts/run-common-tests.sh
@@ -17,6 +17,7 @@ javac --release 21 -cp "$BUILD" -d "$BUILD" "${TEST_SOURCES[@]}"
 
 java -cp "$BUILD" com.chedidandrew.emeraldstandard.core.VanillaConstructionRegressionTest
 java -cp "$BUILD" com.chedidandrew.emeraldstandard.core.DistrictGrowthReviewRegressionTest
+java -cp "$BUILD" com.chedidandrew.emeraldstandard.core.CompactVillageGrowthRegressionTest
 java -cp "$BUILD" com.chedidandrew.emeraldstandard.core.VillageTerritoryRegressionTest
 java -cp "$BUILD" com.chedidandrew.emeraldstandard.client.ExploredTerrainStoreRegressionTest
 java -cp "$BUILD" com.chedidandrew.emeraldstandard.client.NewspaperPagingRegressionTest